backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
python-chess>=1.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
